Added an abstraction over ResizeListener to make it easier to double in tests
To make it easier to fake or mock the resize listener while testing
components that use it, I have introduced a IResizeListener with the
relevant public methods.

This is a breaking change because users will have to switch from injecting
ResizeListener to IResizeListener, since ResizeListener has also been
marked internal. This makes it easier to change the library in the future,
as users cannot use methods/properties/features they are not intented to use.

I also enabled nullable on the project and resolved a few of the resulting
warnings. However, there are still some left, but didnt want go to deep.

Ps. I was not able to get the demo apps to work du to JS errors,
neither before nor after my changes. Perhaps I am missing some build
tools for the JavaScript?