Update readme with BlazorSize info and Telerik promotion
Updated readme.md to include a detailed description of the BlazorSize library, its purpose, and usage for adaptive rendering in Blazor applications. Added a link to the documentation for more information. Included a horizontal rule for better visual organization. Added information about a similar component in Telerik UI for Blazor, along with a promotional image and link to encourage users to try Telerik's offering. Added a note for existing Telerik customers, advising them to use the TelerikMediaQuery component and providing a link to try Telerik UI for Blazor for free.